refactoring: extract part of functionality into `BasicUpdaterPanel` class #7199
